galaxyscape-x-full/requirements.txt
Flask==3.0.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
shap==0.44.0
umap-learn==0.5.5
networkx==3.2.1
karateclub==1.3.0
sqlalchemy==2.0.25
snowflake-connector-python==3.8.0
kafka-python==2.0.2
torch==2.1.2
torchvision==0.16.2
Werkzeug==3.0.1
requests==2.31.0
yfinance==0.2.28
scipy==1.12.0
matplotlib==3.8.2
seaborn==0.13.0
tslearn==0.6.2

